Content Analysis Platform/requirements.txt
matplotlib
pandas
numpy
streamlit==1.40.2
google-cloud-storage==2.16.0
google-cloud-bigquery==3.23.1
db-dtypes==1.2.0
plotly==5.24.1
streamlit-extras
wordcloud